Move types into groups
