+++ src/Language/Mimsa/Project/Versions.hs
@@ -11,13 +11,12 @@ import Language.Mimsa.Project.Usages
 import Language.Mimsa.Types.Error
 import Language.Mimsa.Types.Expr
 import Language.Mimsa.Types.ExprHash
+import Language.Mimsa.Types.Identifiers
 import Language.Mimsa.Types.MonoType
-import Language.Mimsa.Types.Name
 import Language.Mimsa.Types.Project
 import Language.Mimsa.Types.Store
 import Language.Mimsa.Types.StoreExpression
 import Language.Mimsa.Types.Usage
-import Language.Mimsa.Types.Variable
 import Language.Mimsa.Types.VersionedMap
 
 -- find which versions of a given binding are in use
