+++ src/Language/Mimsa/Typechecker/DataTypes.hs
@@ -8,9 +8,9 @@ where
 
 import Data.Map (Map)
 import qualified Data.Map as M
-import Language.Mimsa.Types.Construct
 import Language.Mimsa.Types.DataType
 import Language.Mimsa.Types.Environment
+import Language.Mimsa.Types.Identifiers
 import Language.Mimsa.Types.MonoType
 
 defaultEnv :: Environment
