+++ src/Language/Mimsa/Typechecker/Generalise.hs
@@ -2,10 +2,10 @@ module Language.Mimsa.Typechecker.Generalise where
 
 import Data.List
 import qualified Data.Map as M
+import Language.Mimsa.Types.Identifiers
 import Language.Mimsa.Types.MonoType
 import Language.Mimsa.Types.Scheme
 import Language.Mimsa.Types.Substitutions
-import Language.Mimsa.Types.Variable
 
 freeVars :: MonoType -> [Variable]
 freeVars (MTVar b) = [b]
