+++ src/Language/Mimsa/Types.hs
@@ -1,12 +1,11 @@
 module Language.Mimsa.Types
-  ( module Language.Mimsa.Types.Name,
-    module Language.Mimsa.Types.TypeConstructor,
+  ( module Language.Mimsa.Types.TypeConstructor,
+    module Language.Mimsa.Types.Identifiers,
     module Language.Mimsa.Types.ExprHash,
     module Language.Mimsa.Types.Usage,
     module Language.Mimsa.Types.Expr,
     module Language.Mimsa.Types.Literal,
     module Language.Mimsa.Types.Store,
-    module Language.Mimsa.Types.TypeError,
     module Language.Mimsa.Types.MonoType,
     module Language.Mimsa.Types.Scheme,
     module Language.Mimsa.Types.ForeignFunc,
@@ -14,10 +13,6 @@ module Language.Mimsa.Types
     module Language.Mimsa.Types.ResolvedDeps,
     module Language.Mimsa.Types.Environment,
     module Language.Mimsa.Types.Error,
-    module Language.Mimsa.Types.ResolverError,
-    module Language.Mimsa.Types.InterpreterError,
-    module Language.Mimsa.Types.Variable,
-    module Language.Mimsa.Types.Construct,
     module Language.Mimsa.Types.Scope,
     module Language.Mimsa.Types.Swaps,
     module Language.Mimsa.Types.UniVar,
@@ -26,8 +21,6 @@ module Language.Mimsa.Types
     module Language.Mimsa.Types.Bindings,
     module Language.Mimsa.Types.StoreExpression,
     module Language.Mimsa.Types.Project,
-    module Language.Mimsa.Types.TypeName,
-    module Language.Mimsa.Types.FuncName,
     module Language.Mimsa.Types.StringType,
     module Language.Mimsa.Types.DataType,
     module Language.Mimsa.Types.TypeBindings,
@@ -36,22 +29,18 @@ module Language.Mimsa.Types
 where
 
 import Language.Mimsa.Types.Bindings
-import Language.Mimsa.Types.Construct
 import Language.Mimsa.Types.DataType
 import Language.Mimsa.Types.Environment
 import Language.Mimsa.Types.Error
 import Language.Mimsa.Types.Expr
 import Language.Mimsa.Types.ExprHash
 import Language.Mimsa.Types.ForeignFunc
-import Language.Mimsa.Types.FuncName
-import Language.Mimsa.Types.InterpreterError
+import Language.Mimsa.Types.Identifiers
 import Language.Mimsa.Types.Library
 import Language.Mimsa.Types.Literal
 import Language.Mimsa.Types.MonoType
-import Language.Mimsa.Types.Name
 import Language.Mimsa.Types.Project
 import Language.Mimsa.Types.ResolvedDeps
-import Language.Mimsa.Types.ResolverError
 import Language.Mimsa.Types.Scheme
 import Language.Mimsa.Types.Scope
 import Language.Mimsa.Types.Store
@@ -62,11 +51,8 @@ import Language.Mimsa.Types.Swaps
 import Language.Mimsa.Types.Tui
 import Language.Mimsa.Types.TypeBindings
 import Language.Mimsa.Types.TypeConstructor
-import Language.Mimsa.Types.TypeError
-import Language.Mimsa.Types.TypeName
 import Language.Mimsa.Types.UniVar
 import Language.Mimsa.Types.Usage
-import Language.Mimsa.Types.Variable
 import Language.Mimsa.Types.VersionedMap
 
 ------
