+++ src/Language/Mimsa/Types/Bindings.hs
@@ -10,7 +10,7 @@ import qualified Data.Map as M
 import qualified Data.Text as T
 import Language.Mimsa.Printer
 import Language.Mimsa.Types.ExprHash
-import Language.Mimsa.Types.Name
+import Language.Mimsa.Types.Identifiers
 
 -- a list of names to hashes
 newtype Bindings = Bindings {getBindings :: Map Name ExprHash}
