+++ src/Language/Mimsa/Types/Error/ResolverError.hs
@@ -1,12 +1,12 @@
 {-# LANGUAGE OverloadedStrings #-}
 
-module Language.Mimsa.Types.ResolverError where
+module Language.Mimsa.Types.Error.ResolverError where
 
 import qualified Data.Map as M
 import qualified Data.Text as T
 import Language.Mimsa.Printer
 import Language.Mimsa.Types.Bindings
-import Language.Mimsa.Types.Name
+import Language.Mimsa.Types.Identifiers
 
 data ResolverError = MissingBinding Name Bindings
   deriving (Eq, Ord, Show)
