+++ src/Language/Mimsa/Types/Error/TypeError.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE OverloadedStrings #-}
 
-module Language.Mimsa.Types.TypeError
+module Language.Mimsa.Types.Error.TypeError
   ( TypeError (..),
   )
 where
@@ -13,14 +13,12 @@ import Data.Set (Set)
 import qualified Data.Set as S
 import Data.Text.Prettyprint.Doc
 import Language.Mimsa.Printer
-import Language.Mimsa.Types.Construct
 import Language.Mimsa.Types.DataType
 import Language.Mimsa.Types.Environment
 import Language.Mimsa.Types.Expr
+import Language.Mimsa.Types.Identifiers
 import Language.Mimsa.Types.MonoType
-import Language.Mimsa.Types.Name
 import Language.Mimsa.Types.Swaps
-import Language.Mimsa.Types.Variable
 
 data TypeError
   = UnknownTypeError
