+++ src/Language/Mimsa/Types/ForeignFunc.hs
@@ -1,8 +1,8 @@
 module Language.Mimsa.Types.ForeignFunc where
 
 import Language.Mimsa.Types.Expr
+import Language.Mimsa.Types.Identifiers
 import Language.Mimsa.Types.MonoType
-import Language.Mimsa.Types.Variable
 
 type FFExpr = Expr Variable
