+++ src/Language/Mimsa/Types/Identifiers/TypeName.hs
@@ -1,14 +1,14 @@
 {-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE DeriveGeneric #-}
 
-module Language.Mimsa.Types.TypeName where
+module Language.Mimsa.Types.Identifiers.TypeName where
 
 import qualified Data.Aeson as JSON
 import Data.Text.Prettyprint.Doc
 import GHC.Generics
 import Language.Mimsa.Printer
-import Language.Mimsa.Types.Construct
-import Language.Mimsa.Types.Name
+import Language.Mimsa.Types.Identifiers.Construct
+import Language.Mimsa.Types.Identifiers.Name
 
 data TypeName = ConsName Construct [TypeName] | VarName Name
   deriving (Eq, Ord, Show, Generic, JSON.ToJSON, JSON.FromJSON)
