+++ src/Language/Mimsa/Types/Library.hs
@@ -2,7 +2,7 @@ module Language.Mimsa.Types.Library where
 
 import Data.Map (Map)
 import Language.Mimsa.Types.ForeignFunc
-import Language.Mimsa.Types.FuncName
+import Language.Mimsa.Types.Identifiers
 
 -- our built-in functions for doing IO things etc
 -- statically defined and made available in all computations for now
