+++ src/Language/Mimsa/Types/Scheme.hs
@@ -4,8 +4,8 @@ module Language.Mimsa.Types.Scheme where
 
 import qualified Data.Text as T
 import Language.Mimsa.Printer
+import Language.Mimsa.Types.Identifiers
 import Language.Mimsa.Types.MonoType
-import Language.Mimsa.Types.Variable
 
 data Scheme = Scheme [Variable] MonoType
   deriving (Eq, Ord, Show)
