+++ src/Language/Mimsa/Types/Tui.hs
@@ -3,8 +3,8 @@ module Language.Mimsa.Types.Tui where
 import qualified Brick.Widgets.List as L
 import Data.List.NonEmpty (NonEmpty)
 import Language.Mimsa.Types.Expr
+import Language.Mimsa.Types.Identifiers
 import Language.Mimsa.Types.MonoType
-import Language.Mimsa.Types.Name
 import Language.Mimsa.Types.Project
 import Language.Mimsa.Types.ResolvedDeps
