+++ src/Language/Mimsa/Types/TypeBindings.hs
@@ -10,7 +10,7 @@ import qualified Data.Map as M
 import qualified Data.Text as T
 import Language.Mimsa.Printer
 import Language.Mimsa.Types.ExprHash
-import Language.Mimsa.Types.Name
+import Language.Mimsa.Types.Identifiers
 
 -- a list of names to hashes
 newtype TypeBindings = TypeBindings {getTypeBindings :: Map Name ExprHash}
