+++ src/Language/Mimsa/Types/Usage.hs
@@ -5,7 +5,7 @@ module Language.Mimsa.Types.Usage where
 import qualified Data.Text as T
 import Language.Mimsa.Printer
 import Language.Mimsa.Types.ExprHash
-import Language.Mimsa.Types.Name
+import Language.Mimsa.Types.Identifiers
 
 data Usage
   = Transient Name ExprHash
